🔨 [projects] Move function `*commitmessage` to `repository`
